compares int with int instead of std;:size_t

+++ weaponlist.cpp
@@ -21,7 +21,7 @@ int weapon_name() {
     // Gets weapon input from user. 
     std::cout << "Enter the weapon you have equipped: ";
     std::string weapon{input_handler_str()};
-    for (int elements{0}; elements < std::size(bows); ++elements) {
+    for (int elements{0}; elements < (int) std::size(bows); ++elements) {
         if (weapon == bows[elements]) {
             if ((weapon[0] == 'P') || (weapon[0] == 'p')) {
                 std::cout << "Your weapon is: Prototype Crescent" << std::endl;
